Simplified the StorageAdapter contract
