+++ src/contracts/storage/storage-adapter.contract.ts
@@ -2,7 +2,6 @@
  * @module Storage
  */
 
-import type { GetOrAddResult } from "@/_shared/types";
 import {
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     type StorageError,
@@ -19,15 +18,6 @@ import {
  * @group Contracts
  */
 export type IStorageAdapter<TType> = {
-    /**
-     * Returns true for the keys that are found otherwise false will be returned.
-     * @throws {StorageError} {@link StorageError}
-     * @throws {UnexpectedStorageError} {@link UnexpectedStorageError}
-     */
-    existsMany?<TKeys extends string>(
-        keys: TKeys[],
-    ): PromiseLike<Record<TKeys, boolean>>;
-
     /**
      * Returns the value for the keys that are found otherwise null will be returned.
      * @throws {StorageError} {@link StorageError}
@@ -60,7 +50,7 @@ export type IStorageAdapter<TType> = {
      * @throws {StorageError} {@link StorageError}
      * @throws {UnexpectedStorageError} {@link UnexpectedStorageError}
      */
-    putMany?<TValues extends TType, TKeys extends string>(
+    putMany<TValues extends TType, TKeys extends string>(
         values: Record<TKeys, TValues>,
     ): PromiseLike<Record<TKeys, boolean>>;
 
@@ -74,31 +64,12 @@ export type IStorageAdapter<TType> = {
     ): PromiseLike<Record<TKeys, boolean>>;
 
     /**
-     * If the key is found the value be returned and key will be removed otherwise null will be returned.
-     * @throws {StorageError} {@link StorageError}
-     * @throws {UnexpectedStorageError} {@link UnexpectedStorageError}
-     */
-    getAndRemove?<TValue extends TType>(
-        key: string,
-    ): PromiseLike<TValue | null>;
-
-    /**
-     * If the key is found the value be returned otherwise valueToAdd will be added and returned.
-     * @throws {StorageError} {@link StorageError}
-     * @throws {UnexpectedStorageError} {@link UnexpectedStorageError}
-     */
-    getOrAdd?<TValue extends TType, TExtended extends TType>(
-        key: string,
-        valueToAdd: TExtended,
-    ): PromiseLike<GetOrAddResult<TValue | TExtended>>;
-
-    /**
      * Will increment the existing key with value if found otherwise nonthing occurs. Returns true if key exists otherwise false will be returned.
      * @throws {StorageError} {@link StorageError}
      * @throws {UnexpectedStorageError} {@link UnexpectedStorageError}
      * @throws {TypeStorageError} {@link TypeStorageError}
      */
-    increment?(key: string, value: number): PromiseLike<boolean>;
+    increment(key: string, value: number): PromiseLike<boolean>;
 
     /**
      * @throws {StorageError} {@link StorageError}
